Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit==0.19.1",
    "livekit-agents==0.12.11",
    "livekit-plugins-openai==0.10.17",
    "python-dotenv>=1.1.1",
]
